revisao

+++ calculadora-switch/test/CalculadoraTest.java
@@ -1,32 +1,45 @@
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 public class CalculadoraTest {
     @Test
     void deveRetornarASomaDeDoisComTres() {
-        assertEquals(5, Calculadora.somar(2, 3));
+        assertEquals(BigDecimal.valueOf(5),
+                Calculadora.somar(BigDecimal.valueOf(2), BigDecimal.valueOf(3))
+        );
     }
 
     @Test
     void deveRetornarASubtracaoDeCincoComTres() {
-        assertEquals(2, Calculadora.subtrair(5, 3));
+        assertEquals(BigDecimal.valueOf(2),
+                Calculadora.subtrair(BigDecimal.valueOf(5), BigDecimal.valueOf(3))
+        );
     }
 
     @Test
     void deveRetornarADivisaoDeOitoComDois() {
-        assertEquals(4, Calculadora.dividir(8, 2));
+        assertEquals(BigDecimal.valueOf(4),
+                Calculadora.dividir(BigDecimal.valueOf(8), BigDecimal.valueOf(2))
+        );
     }
 
     @Test
     void deveRetornarAMultiplicacaoDeQuatroComCinco() {
-        assertEquals(20, Calculadora.multiplicar(4, 5));
+        assertEquals(BigDecimal.valueOf(20),
+                Calculadora.multiplicar(BigDecimal.valueOf(4), BigDecimal.valueOf(5)));
     }
 
     @Test
     void deveRetornarAPotenciacao() {
-        assertEquals(16, Calculadora.potencia(4, 2));
-        assertEquals(125, Calculadora.potencia(5, 3));
+        assertEquals(BigDecimal.valueOf(16),
+                Calculadora.potencia(BigDecimal.valueOf(4), BigDecimal.valueOf(2))
+        );
+        assertEquals(BigDecimal.valueOf(125),
+                Calculadora.potencia(BigDecimal.valueOf(5), BigDecimal.valueOf(3))
+        );
     }
 
 }
